// editor-app/src/entities/EditorBlocks/MjmlButton/MjmlButton.types.ts
export interface MjmlButtonProps {
	text: string
	url: string
	color: string
	backgroundColor: string
	borderRadius: string
	width: string
	height: string
	align: 'left' | 'center' | 'right'
	paddingTop: string
	paddingRight: string
	paddingBottom: string
	paddingLeft: string
	[key: string]: unknown
}